Sphere metric

+++ src/term.h
@@ -13,7 +13,7 @@ struct term {
 };
 
 struct term* termalloc(struct elt *, int, double);
-struct term* pairproduct(struct elt *, struct elt *, int);
+struct term* pairproduct(struct elt *, struct elt *, double);
 struct term* termpermute(struct term *, const int *);
 struct term* termperms(struct term *, int);
 void printterm(struct term *);
